Move delta window functions to package functions.windowing.delta
