+++ flink-staging/flink-streaming/flink-streaming-core/src/test/java/org/apache/flink/streaming/api/windowing/deltafunction/EuclideanDistanceTest.java
@@ -19,6 +19,7 @@ package org.apache.flink.streaming.api.windowing.deltafunction;
 
 import static org.junit.Assert.*;
 
+import org.apache.flink.streaming.api.functions.windowing.delta.EuclideanDistance;
 import org.junit.Test;
 
 public class EuclideanDistanceTest {
